:tada: 0.1.0
Upgrade to 1.13.8, rebuilt from scratch since it was unclear how well
used the existing project was but it has had no changes in a year now.

Time for a spring clean.

Bumped a minor version in case we actually have any users and they
accidentally automatically update to this release, which is _completely_
different.